refactor : exception 위치 library 디렉토리로 옮김
issue : #129 #130
