+++ src/main/java/keeper/project/homepage/exception/library/CustomBookOverTheMaxException.java
@@ -1,4 +1,4 @@
-package keeper.project.homepage.exception;
+package keeper.project.homepage.exception.library;
 
 public class CustomBookOverTheMaxException extends RuntimeException {
